Send WS protocol Version
The `Sec-WebSocket-Protocol` header field is used in the WebSocket 
opening handshake.  It is sent from the client to the server and back
from the server to the client to confirm the subprotocol of the
connection.  This enables scripts to both select a subprotocol and be
sure that the server agreed to serve that subprotocol.

https://tools.ietf.org/html/rfc6455#section-11.3.4

The idea is to prevent requiring backwards compatibility if it happens
that we want to change the payload we’re sending to the server