add support for trace(method, rtcpeerconnection, ...)
Summary:
Ref T27155
adds support for calling trace with a RTCPeerConnection as second argument instead of the id. This will resolve the id according to the "private" convention for storing the rtcstats id.
This allows external calls to trace() to be associated with a certain peerconnection

Test Plan: - call trace with a peerconnection, observe id being traced

Reviewers: gustavo

Reviewed By: gustavo

Maniphest Tasks: T27155

Differential Revision: https://lifeonair.phacility.com/D26230